Moving some of the aliases under PryByebug
Some of the aliases don’t work with the Pry gem, they specifically come
as part of other gems like
PryByeBug(https://github.com/deivid-rodriguez/pry-byebug). To ensure
that the other aliases are loaded properly(i.e., without raising an
error as part of .pryrc) when say using only the pry gem, we ensure
that the prybyebug aliases are defined when the prybyebug gem is
present in the list of gems available to use for a given development
environment.