Add an alias for reloading an object
Misc change: Rename alias for require. One of the reasons for renaming the "require" alias is I wanted to make use of the alias(or method name as you see based on the way it is defined in my .pryrc) r for "reload!" in the context of reloading an object whenever you've made changes to your Rails models and to ensure that the object picks up those latest changes.